refactor(desk-tool): improve `PaneRouterProvider`
