+++ packages/@sanity/desk-tool/src/contexts/paneRouter/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * @internal
+ */
+export const exclusiveParams = ['view', 'since', 'rev']
