+++ packages/@sanity/desk-tool/src/contexts/paneRouter/index.ts
@@ -1,0 +1,5 @@
+export * from './BackLink'
+export {exclusiveParams} from './constants'
+export * from './PaneRouterProvider'
+export * from './types'
+export * from './usePaneRouter'
