+++ packages/@sanity/desk-tool/src/contexts/paneRouter/usePaneRouter.ts
@@ -1,0 +1,10 @@
+import {useContext} from 'react'
+import {PaneRouterContext} from './PaneRouterContext'
+import {PaneRouterContextValue} from './types'
+
+/**
+ * @public
+ */
+export function usePaneRouter(): PaneRouterContextValue {
+  return useContext(PaneRouterContext)
+}
